rename to react_native_tests and update buildTooldVersion to 28.0.3
