+++ .gitignore
@@ -56,4 +56,4 @@ buck-out/
 *.jsbundle
 
 # test coverage
-coverage/*
+coverage/
