+++ README.md
@@ -1,2 +1,20 @@
 # react_native_tests
-simple react native test example with redux
+simple react native test examples with redux
+
+
+### to run tests:
+- `yarn install` note: I chose yarn over npm. Either will work but React Native sometimes has issues if you have both yarn.lock and package-lock.json. If you choose npm, delete yarn.lock first.
+- `test` OR `yarn testcov` (`test --coverage`)
+
+
+#### TODO:
+- [ ] add actions and action tests
+- [ ] add connected components and test
+- [ ] add `reselect` library then implement and test
+- [ ] add fetchMock tests (and create something to pretend to fetch)
+- [ ] add smoke tests
+- [ ] and mount tests
+
+
+
+##### PR's welcome to develop branch
