+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.react_native_method_test",
+    package = "com.react_native_tests",
 )
 
 android_resource(
     name = "res",
-    package = "com.react_native_method_test",
+    package = "com.react_native_tests",
     res = "src/main/res",
 )
