+++ android/app/src/main/java/com/react_native_tests/MainApplication.java
@@ -1,4 +1,4 @@
-package com.react_native_method_test;
+package com.react_native_tests;
 
 import android.app.Application;
