+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">react_native_method_test</string>
+    <string name="app_name">react_native_tests</string>
 </resources>
