+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'react_native_method_test'
+rootProject.name = 'react_native_tests'
 
 include ':app'
