+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "react_native_method_test",
-  "displayName": "react_native_method_test"
+  "name": "react_native_tests",
+  "displayName": "react_native_tests"
 }
