+++ ios/react_native_tests-tvOS/Info.plist
