+++ ios/react_native_tests-tvOSTests/Info.plist
