+++ ios/react_native_tests.xcodeproj/xcshareddata/xcschemes/react_native_tests-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "react_native_method_test-tvOS.app"
-               BlueprintName = "react_native_method_test-tvOS"
-               ReferencedContainer = "container:react_native_method_test.xcodeproj">
+               BuildableName = "react_native_tests-tvOS.app"
+               BlueprintName = "react_native_tests-tvOS"
+               ReferencedContainer = "container:react_native_tests.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "react_native_method_test-tvOSTests.xctest"
-               BlueprintName = "react_native_method_test-tvOSTests"
-               ReferencedContainer = "container:react_native_method_test.xcodeproj">
+               BuildableName = "react_native_tests-tvOSTests.xctest"
+               BlueprintName = "react_native_tests-tvOSTests"
+               ReferencedContainer = "container:react_native_tests.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "react_native_method_test-tvOSTests.xctest"
-               BlueprintName = "react_native_method_test-tvOSTests"
-               ReferencedContainer = "container:react_native_method_test.xcodeproj">
+               BuildableName = "react_native_tests-tvOSTests.xctest"
+               BlueprintName = "react_native_tests-tvOSTests"
+               ReferencedContainer = "container:react_native_tests.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "react_native_method_test-tvOS.app"
-            BlueprintName = "react_native_method_test-tvOS"
-            ReferencedContainer = "container:react_native_method_test.xcodeproj">
+            BuildableName = "react_native_tests-tvOS.app"
+            BlueprintName = "react_native_tests-tvOS"
+            ReferencedContainer = "container:react_native_tests.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "react_native_method_test-tvOS.app"
-            BlueprintName = "react_native_method_test-tvOS"
-            ReferencedContainer = "container:react_native_method_test.xcodeproj">
+            BuildableName = "react_native_tests-tvOS.app"
+            BlueprintName = "react_native_tests-tvOS"
+            ReferencedContainer = "container:react_native_tests.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "react_native_method_test-tvOS.app"
-            BlueprintName = "react_native_method_test-tvOS"
-            ReferencedContainer = "container:react_native_method_test.xcodeproj">
+            BuildableName = "react_native_tests-tvOS.app"
+            BlueprintName = "react_native_tests-tvOS"
+            ReferencedContainer = "container:react_native_tests.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
