+++ ios/react_native_tests/Images.xcassets/AppIcon.appiconset/Contents.json
