+++ ios/react_native_tests/Images.xcassets/Contents.json
