+++ ios/react_native_testsTests/Info.plist
