+++ ios/react_native_testsTests/react_native_testsTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface react_native_method_testTests : XCTestCase
+@interface react_native_testsTests : XCTestCase
 
 @end
 
-@implementation react_native_method_testTests
+@implementation react_native_testsTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
