+++ package.json
@@ -1,10 +1,11 @@
 {
-  "name": "react_native_method_test",
+  "name": "react_native_tests",
   "version": "0.0.1",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "test": "jest",
+    "testcov": "jest --coverage"
   },
   "dependencies": {
     "prop-types": "^15.7.2",
